requirements.txt
# Core Framework
streamlit==1.33.0

# PDF Handling
PyPDF2==3.0.1

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3

# Hugging Face Transformers
transformers==4.40.1
sentence-transformers==2.2.2
torch==2.2.1
numpy==1.26.4

# FAISS for Semantic Search
faiss-cpu==1.7.4

# File Utilities
tqdm==4.66.2

# Environment
python-dotenv==1.0.1
